test/changhee/obstacle_v2/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0.0
pyyaml>=6.0
